# Conflicts:
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pub/src/main/java/nts/uk/ctx/bs/employee/pub/workplace/SyWorkplacePub.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pubimp/src/main/java/nts/uk/ctx/bs/employee/pubimp/workplace/WorkplacePubImp.java
#	nts.uk/uk.com/com.web/nts.uk.com.web/src/main/webapp/view/cmm/011/a/index.xhtml

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/divergence/time/DivergenceTimeSaveCommand.java
@@ -40,7 +40,7 @@ public class DivergenceTimeSaveCommand implements DivergenceTimeGetMemento {
 	private boolean reasonSelect;
 
 	/** Attendance Item list. */
-	private List<Double> attendanceId;
+	private List<Integer> attendanceId;
 
 	/**
 	 * Instantiates a new divergence time save command.
@@ -90,7 +90,7 @@ public class DivergenceTimeSaveCommand implements DivergenceTimeGetMemento {
 	 * @see nts.uk.ctx.at.record.dom.divergence.time.DivergenceTimeGetMemento#getTargetItems()
 	 */
 	@Override
-	public List<Double> getTargetItems() {
+	public List<Integer> getTargetItems() {
 		return this.attendanceId;
 
 	}
